refactor(qdrant): consolidate Qdrant client into knowledgebase module and update settings structure
