--- zammad-ai/app/qdrant/__init__.py
@@ -1,3 +1,0 @@
-from .qdrant import QdrantKBClient, QdrantKBError
-
-__all__: list[str] = ["QdrantKBClient", "QdrantKBError"]
